% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/humanparser-package.r
\docType{package}
\name{humanparser}
\alias{humanparser}
\alias{humanparser-package}
\title{A package to parse human names}
\description{
A package to parse human names
}
\author{
Bob Rudis (@hrbrmstr)
}

