% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/humanparser.r
\name{parse_name}
\alias{parse_name}
\title{Parse a "human" name into components}
\usage{
parse_name(full_name)
}
\arguments{
\item{full_name}{full name to parse}
}
\value{
\code{list} with name parsed
}
\description{
Parse a "human" name into components
}
\examples{
parse_name("John Smith Jr.")
}

