% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/humanparser.r
\name{parse_names}
\alias{parse_names}
\title{Parse a character vector of names into a data frame}
\usage{
parse_names(full_names)
}
\arguments{
\item{full_names}{character vector of names}
}
\value{
\code{data.frame} with names parsed
}
\description{
Parse a character vector of names into a data frame
}
\examples{
full_names <- c("David Regan", "Izaque Iuzuru Nagata",
                "Christian Schmit de la Breli", "Peter Doyle", "Hans R.Bruetsch",
                "Marcus Reichel", "Per-Axel Koch", "Louis Van der Walt",
                "Mario Adamek", "Ugur Tozsekerli", "Judit Ludvai" )
parse_names(full_names)
}

